[DAS -208] - Added Regions enum, Removed Inline models and Accumulated Docs team changes

+++ oss/source/Model/ObjectFullDetails.gen.cs
@@ -33,7 +33,7 @@ using Newtonsoft.Json.Converters;
 namespace Autodesk.Oss.Model
 {
     /// <summary>
-    /// Object Details json response
+    /// Represents detailed information about an object within a bucket.
     /// </summary>
     [DataContract]
     public partial class ObjectFullDetails 
@@ -46,79 +46,101 @@ namespace Autodesk.Oss.Model
         }
         
         /// <summary>
-        /// Bucket key
+        ///The bucket key of the bucket that contains the object.
         /// </summary>
-        /// <value>Bucket key</value>
+        /// <value>
+        ///The bucket key of the bucket that contains the object.
+        /// </value>
         [DataMember(Name="bucketKey", EmitDefaultValue=false)]
         public string BucketKey { get; set; }
 
         /// <summary>
-        /// Object URN
+        ///An identifier (URN) that uniquely and persistently identifies the object.
         /// </summary>
-        /// <value>Object URN</value>
+        /// <value>
+        ///An identifier (URN) that uniquely and persistently identifies the object.
+        /// </value>
         [DataMember(Name="objectId", EmitDefaultValue=false)]
         public string ObjectId { get; set; }
 
         /// <summary>
-        /// Object name
+        ///A URL-encoded human friendly name to identify the object.
         /// </summary>
-        /// <value>Object name</value>
+        /// <value>
+        ///A URL-encoded human friendly name to identify the object.
+        /// </value>
         [DataMember(Name="objectKey", EmitDefaultValue=false)]
         public string ObjectKey { get; set; }
 
         /// <summary>
-        /// Object SHA1
+        ///A hash value computed from the data of the object.
         /// </summary>
-        /// <value>Object SHA1</value>
+        /// <value>
+        ///A hash value computed from the data of the object.
+        /// </value>
         [DataMember(Name="sha1", EmitDefaultValue=false)]
         public byte[] Sha1 { get; set; }
 
         /// <summary>
-        /// Object size
+        ///The total amount of storage space occupied by the object, in bytes.
         /// </summary>
-        /// <value>Object size</value>
+        /// <value>
+        ///The total amount of storage space occupied by the object, in bytes.
+        /// </value>
         [DataMember(Name="size", EmitDefaultValue=false)]
-        public long? Size { get; set; }
+        public int? Size { get; set; }
 
         /// <summary>
-        /// Object content-type
+        ///The format of the data stored within the object, expressed as a MIME type.
         /// </summary>
-        /// <value>Object content-type</value>
+        /// <value>
+        ///The format of the data stored within the object, expressed as a MIME type.
+        /// </value>
         [DataMember(Name="contentType", EmitDefaultValue=false)]
         public string ContentType { get; set; }
 
         /// <summary>
-        /// URL to download the object
+        ///A URL that points to the actual location of the object.
         /// </summary>
-        /// <value>URL to download the object</value>
+        /// <value>
+        ///A URL that points to the actual location of the object.
+        /// </value>
         [DataMember(Name="location", EmitDefaultValue=false)]
         public string Location { get; set; }
 
         /// <summary>
-        /// Epoch timestamp when the object was created. Only provided if the user requests this to be included in the query parameters.
+        ///The time the object was created, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
         /// </summary>
-        /// <value>Epoch timestamp when the object was created. Only provided if the user requests this to be included in the query parameters.</value>
+        /// <value>
+        ///The time the object was created, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
+        /// </value>
         [DataMember(Name="createdDate", EmitDefaultValue=false)]
         public long? CreatedDate { get; set; }
 
         /// <summary>
-        /// Epoch timestamp when the object was last accessed. Only provided if the user requests this to be included in the query parameters.
+        ///The time the object was last accessed, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
         /// </summary>
-        /// <value>Epoch timestamp when the object was last accessed. Only provided if the user requests this to be included in the query parameters.</value>
+        /// <value>
+        ///The time the object was last accessed, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
+        /// </value>
         [DataMember(Name="lastAccessedDate", EmitDefaultValue=false)]
         public long? LastAccessedDate { get; set; }
 
         /// <summary>
-        /// Epoch timestamp when the object was last modified. Only provided if the user requests this to be included in the query parameters.
+        ///The time the object was most recently modified, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
         /// </summary>
-        /// <value>Epoch timestamp when the object was last modified. Only provided if the user requests this to be included in the query parameters.</value>
+        /// <value>
+        ///The time the object was most recently modified, represented as a Unix timestamp. Only returned if explicitly requested using the `with` query string parameter.
+        /// </value>
         [DataMember(Name="lastModifiedDate", EmitDefaultValue=false)]
         public long? LastModifiedDate { get; set; }
 
         /// <summary>
-        /// When specified, the response will include the user defined metadata in a JSON field named userDefinedMetadata, if available. The user defined metadata can be set on file uploading.
+        ///Any custom metadata, if available. Only returned if explicitly requested for using the `with` query string parameter.
         /// </summary>
-        /// <value>When specified, the response will include the user defined metadata in a JSON field named userDefinedMetadata, if available. The user defined metadata can be set on file uploading.</value>
+        /// <value>
+        ///Any custom metadata, if available. Only returned if explicitly requested for using the `with` query string parameter.
+        /// </value>
         [DataMember(Name="userDefinedMetadata", EmitDefaultValue=false)]
         public string UserDefinedMetadata { get; set; }
